+++ sdk/js/src/solana/wormhole/coder/instruction.ts
@@ -1,114 +1,10 @@
 import { Idl, Instruction, InstructionCoder } from "@project-serum/anchor";
+import { bs58 } from "@project-serum/anchor/dist/cjs/utils/bytes";
 import { Layout } from "buffer-layout";
 import { camelCase } from "lodash";
-import { IdlField, IdlStateMethod, IdlTypeDef } from "../../anchor";
+import { IdlField, IdlStateMethod } from "../../anchor";
 import * as borsh from "@coral-xyz/borsh"
-import { bs58 } from "@project-serum/anchor/dist/cjs/utils/bytes";
-import { toPascalCase } from "@injectivelabs/sdk-ts";
-
-export class IdlCoder {
-  public static fieldLayout(
-    field: { name?: string } & Pick<IdlField, "type">,
-    types?: IdlTypeDef[]
-  ): Layout {
-    const fieldName =
-      field.name !== undefined ? camelCase(field.name) : undefined;
-    switch (field.type) {
-      case "bool": {
-        return borsh.bool(fieldName);
-      }
-      case "u8": {
-        return borsh.u8(fieldName);
-      }
-      case "i8": {
-        return borsh.i8(fieldName);
-      }
-      case "u16": {
-        return borsh.u16(fieldName);
-      }
-      case "i16": {
-        return borsh.i16(fieldName);
-      }
-      case "u32": {
-        return borsh.u32(fieldName);
-      }
-      case "i32": {
-        return borsh.i32(fieldName);
-      }
-      case "f32": {
-        return borsh.f32(fieldName);
-      }
-      case "u64": {
-        return borsh.u64(fieldName);
-      }
-      case "i64": {
-        return borsh.i64(fieldName);
-      }
-      case "f64": {
-        return borsh.f64(fieldName);
-      }
-      case "u128": {
-        return borsh.u128(fieldName);
-      }
-      case "i128": {
-        return borsh.i128(fieldName);
-      }
-      case "u256": {
-        return borsh.u256(fieldName);
-      }
-      case "i256": {
-        return borsh.i256(fieldName);
-      }
-      case "bytes": {
-        return borsh.vecU8(fieldName);
-      }
-      case "string": {
-        return borsh.str(fieldName);
-      }
-      case "publicKey": {
-        return borsh.publicKey(fieldName);
-      }
-      default: {
-        if ("vec" in field.type) {
-          return borsh.vec(
-            IdlCoder.fieldLayout(
-              {
-                name: undefined,
-                type: field.type.vec,
-              },
-              types
-            ),
-            fieldName
-          );
-        } else if ("option" in field.type) {
-          return borsh.option(
-            IdlCoder.fieldLayout(
-              {
-                name: undefined,
-                type: field.type.option,
-              },
-              types
-            ),
-            fieldName
-          );
-        } else if ("array" in field.type) {
-          let arrayTy = field.type.array[0];
-          let arrayLen = field.type.array[1];
-          let innerLayout = IdlCoder.fieldLayout(
-            {
-              name: undefined,
-              type: arrayTy,
-            },
-            types
-          );
-          return borsh.array(innerLayout, arrayLen, fieldName);
-        } else {
-          throw new Error(`Not yet implemented: ${field}`);
-        }
-      }
-    }
-  }
-}
+import { IdlCoder } from "./idl";
   
 export class WormholeInstructionCoder implements InstructionCoder {
 
@@ -156,8 +52,6 @@ export class WormholeInstructionCoder implements InstructionCoder {
     }
     const len = layout.encode(ix, buffer);
     const data = buffer.slice(0, len);
-
-    let discriminator = toPascalCase(ixName);
     
         switch (ixName) {
       case "initialize": {
@@ -220,7 +114,7 @@ export class WormholeInstructionCoder implements InstructionCoder {
 
 /** Solitaire enum of existing the Core Bridge's instructions.
  *
- * https://github.com/certusone/wormhole/blob/dev.v2/solana/bridge/program/src/lib.rs#L92
+ * https://github.com/certusone/wormhole/blob/main/solana/bridge/program/src/lib.rs#L92
  */
 export enum WormholeInstruction {
   Initialize,
